+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # Wasat ChangeLog
 
+## Unreleased
+
+**WiP**
+
+- Added support for generating and storing client certificates.
+  ([#3](https://github.com/davep/wasat/pull/3))
+
 ## v0.0.1
 
 **2026-06-17**